Create population.c
Determine how long it takes for a population to reach a particular size.
Say we have a population of n llamas. Each year, n / 3 new llamas are born, and n / 4 llamas pass away.
For example, if we were to start with n = 1200llamas, then in the first year, 1200 / 3 = 400new llamas would be born and 1200 / 4 = 300llamas would pass away. At the end of that year, we would have 1200 + 400 - 300 = 1300 llamas.


$ ./population
Start size: 100
End size: 200
Years: 9